Time: 0 ms (100.00%) | Memory: 42.1 MB (68.57%) - LeetSync

+++ 50-powx-n/powx-n.java
@@ -16,7 +16,7 @@ class Solution {
 
     }
     public double myPow(double x, int n) {
-        double ans=solve(x,(long)n);
+        double ans=solve(x,n);
         return ans;
     }
 }
